Create TheBlackJackTroll
The Blackjack Troll game was built so that anybody may play it for free. The game is modeled on the classic Blackjack card game, but with a twist. It was made with Matlab, and the game is played by requesting the user to enter specific letters. Blackjack troll is a traditional game with simple rules and is a fun way to spend the time since it mimics the scenario of the troll on the bridge. Because Blackjack is not an easy game to write, making a simple plan and executing a test trial every now and then would be quite beneficial. Overall, the team was thrilled at the end when the game code worked well.